fix: Kill previous exec thread before spawning new
Before this fix, previous exec threads would not be killed between
restarts. This means the process would pile up over time and that is
definitely not a good thing.

However, with this patch, the exec thread sends the Child handles to the
main event loop through a new channel where they get dealt with before
the new exec thread is spawned to take it's place.