All: Change visiblity and remove doc strings
No reason for most `pub` when we aren't a library exporting anything. Refactored to use `pub(crate)` instead.

Doc strings are removed because like the previous commit mentioned, we
don't get a docs.rs link anyway since we are a binary and not a library.
The comments take up a tremendous amount of space so trimming most of
them and removing some of them makes the code base much easier to read.